pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "apilens"
version = "0.1.0"
description = "A Python SDK for tracking and optimizing AI-model API usage"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "apilens"}]
license = "MIT"
repository = "https://github.com/yourusername/apilens"
keywords = ["openai", "api", "tracking", "monitoring", "cost"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
openai = ">=1.0.0"
python-dotenv = ">=1.0.0"
fastapi = { version = ">=0.100.0", optional = true }

[tool.poetry.extras]
api = ["fastapi"]

[tool.poetry.dev-dependencies]
pytest = ">=7.0.0"
